+++ controller/tests/test_constants.py
@@ -160,10 +160,10 @@ from mantarray_desktop_app import WELL_24_INDEX_TO_ADC_AND_CH_INDEX
 from mantarray_desktop_app.constants import ALL_VALID_BARCODE_HEADERS
 from mantarray_desktop_app.constants import BARCODE_HEADERS
 from mantarray_desktop_app.constants import BARCODE_LEN
+from mantarray_desktop_app.constants import MAGNET_TYPE_TO_MM_PER_MT_Z_AXIS_SENSOR_0
 from mantarray_desktop_app.constants import MICRO_TO_BASE_CONVERSION
 from mantarray_desktop_app.constants import NUM_INITIAL_MICROSECONDS_TO_PAD
 from mantarray_desktop_app.constants import PERFOMANCE_LOGGING_PERIOD_SECS
-from mantarray_desktop_app.constants import POST_STIFFNESS_TO_MM_PER_MT_Z_AXIS_SENSOR_0
 from mantarray_desktop_app.constants import RECORDING_SNAPSHOT_DUR_SECS
 from mantarray_desktop_app.constants import SERIAL_COMM_NICKNAME_BYTES_LENGTH
 from mantarray_desktop_app.constants import SERIAL_COMM_SERIAL_NUMBER_BYTES_LENGTH
@@ -592,7 +592,7 @@ def test_calibration_constants():
 
 
 def test_live_view_conversion():
-    assert POST_STIFFNESS_TO_MM_PER_MT_Z_AXIS_SENSOR_0 == {1: 8, 3: 8, 6: 8, 9: 8, 12: 6.896551}
+    assert MAGNET_TYPE_TO_MM_PER_MT_Z_AXIS_SENSOR_0 == {"-1": 1 / 0.114, "-2": 1 / 0.114, "-5": 1 / 0.213}
 
 
 def test_recording_snapshot():